Implement push notification functionality in project
A significant overhaul of the project has been done to implement push notifications. First, the old index.html file was removed and replaced with a new dynamic MVC view, Index.cshtml. A HomeController was added to handle requests to /home/index with two methods GET Index for displaying the page and POST Index for saving client push subscription information from the frontend. In order to persist and manage client subscriptions the ApplicationDbContext, PersistedClient entity and DbContext were created. Furthermore, Migration files were also added to represent the database schema changes.

A new HumidityController was created with a TestNotify endpoint, initiating a push notification to each saved client. The ServiceWorker.js file was expanded to be able to handle 'push' and 'notificationclick' events.

Push notification subscription and permission handling was added in the Index.cshtml file.

Corresponding changes were also made in the appsettings.json, project file and Startup.cs to complete these changes and prepare the application for the new functionality.